requirements.txt
google-genai
streamlit
ipython